Changes to models managers for tickets and members
There are some force unwraps introduced, which should theoretically work as long
as we don't call things without having verified the correctness of the server URL
and without having no team set. That should work since we need to ensure that anyways,
but we need to be careful about it.